Refactors WeightedMaxEffort into a more general Event
To allow for the implementation of a common representation
for events of different movements, WeightedMaxEffort has been
replaced with a event. An enum on Event will determine the kind
of event.

Running and rowing events, which only require a single attempt, will
be represented as a single row with success set to true.

These changes should allow easier development of queries required
for implementing the reporting questions in the spec.