Sketches queries for each visualization
These queries will need to be reorganized and tested,
but this me working out the major queries for the various
parts of the app. Next up roughing in a visualization or to
workout my workflow for using D3 in this app.

I think I will use PunchBox to determine what JS to run on
each page.